requirements.txt
streamlit
crewai
langchain
langchain-groq
python-dotenv
langchain-community
